+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project from version 1.2.0 upwards are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.6.12] – 2024-03-07
+
+### Added
+- LionwebNode get method
+
+### Fixed
+- ECoreNode definition
+
 ## [1.6.11] – 2024-03-07
 
 ### Changed